requirements.txt
requests>=2.28,<3
Pillow>=9.0
urllib3>=1.26,<3
